Note: Use Jupyter Notebook; it is effective.
 Steps to run the code:

  #Step 1

 By executing the following pip instructions in a jupyter notebook environment, you can import all of the aforementioned modules. 
                 pip install numpy
                 pip install pandas 
                 pip install tkinter 
                 pip install sklearn 

 #Step2  

Then execute the section 1 code in a fresh Jupyter notebook cell. The reading, loading, and modification of the data set are all contained in this Section 1 code. also educating the model.  

 #Step 3  

Then execute section 2 code in the newly created cell that is interfaced to take the test and provide the necessary data.      
fill the given fields first (age,area,age)       
         Age should be between 22-19      
         Area should be either Rural or Urban       
        Gender should be Male or Female 
        after filling the details click submit  
note: don't forget to click after immediate filling of details then take the following psychometric test(15 Questions) 
note: Answer the questions with the understanding that a fresher will be taking the test. then close the gui   

Step 4:

 run the section 4 code in new cell  which is having testing part and finding accuracy of the model  

Step 5:

 then Run the next section code that is section 5. which opens a new GUI which will display the predicted emotional intelligence quotient of a person.